feat: utilities
